Add NegativeTests and test_data with general data
 Move negative test to right class
