+++ test/search/tests.py
@@ -1,30 +1,10 @@
 # -*- coding: utf-8 -*-
 
 from test.search.page_objects import *
-import time
+from test_data import TestSearch
 
 
-class PositiveTests(unittest.TestCase):
-    QUESTION_ID_PROGRAMMING = u"184484161"
-    QUESTION_TITLE_PROGRAMMING = u"I hope now it is long enough? You need more symbols, really? " \
-                                 u"WTF How much do you need? Да вы блин серьезно?!"
-    PROG_CATEGORY = u'Программирование'
-    PYTHON_SUBCATEGORY = u'Python'
-    USERNAME = u'Артур Пирожков'
-
-    QUESTION_ID_OTHER = u"182362166"
-    QUESTION_TITLE_OTHER = u"Тестирование длины форм -"
-    OTHER_CATEGORY = u'Другое'
-
-    def search(self, page, search_request):
-        top_bar = page.get_top_bar_form
-        top_bar.search(search_request)
-        top_bar.submit()
-        return page.get_search_results_form
-
-    def accurate_search(self, page, search_request):
-        return self.search(page, '\"' + search_request + '\"')
-
+class PositiveTests(TestSearch):
     def tearDown(self):
         # pass
         self.driver.quit()
@@ -86,7 +66,7 @@ class PositiveTests(unittest.TestCase):
         search_page = SearchPage(self.driver)
         search_page.open()
 
-        self.search(search_page, self.QUESTION_TITLE_PROGRAMMING)
+        self.accurate_search(search_page, self.QUESTION_TITLE_PROGRAMMING)
 
         side_bar = search_page.get_side_bar_form
         side_bar.set_category(self.PROG_CATEGORY)
@@ -95,11 +75,27 @@ class PositiveTests(unittest.TestCase):
         search_results = search_page.get_search_results_form
         self.assertTrue(search_results.check_question_exist(self.QUESTION_ID_PROGRAMMING))
 
+
+class NegativeTests(TestSearch):
+
+    def tearDown(self):
+        # pass
+        self.driver.quit()
+
+    def setUp(self):
+        browser = 'CHROME'
+
+        self.driver = Remote(
+            command_executor='http://127.0.0.1:4444/wd/hub',
+            desired_capabilities=getattr(DesiredCapabilities, browser)
+                .copy()
+        )
+
     def test_wrong_category(self):
         search_page = SearchPage(self.driver)
         search_page.open()
 
-        self.search(search_page, self.QUESTION_TITLE_PROGRAMMING)
+        self.accurate_search(search_page, self.QUESTION_TITLE_PROGRAMMING)
 
         side_bar = search_page.get_side_bar_form
         side_bar.set_category(self.OTHER_CATEGORY)
